Mega-commit, WIP. See commit message body.
I haven't touched Java in a long time, so I was fumbling my way around
the language and libraries. Didn't see much of a point in tracking me
trying to find a solution I was comfortable using.
I tried using Spring Data REST, but implementing the calls which deal
with linked nodes gave me so much trouble that I gave up on it. The
current state does include the library, but I'll soon remove it.

A summary of what's included in this commit is as follows:
* Add permission and permission node classes
* Implement GET, POST, PUT, DELETE for permission node API
* Implement GET for permission API